+++ config/easy/MakeAnAtariFont/code/CGCode.js
@@ -1,0 +1,6 @@
+const aWord = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('that word in Atari Font');